bugfix: check of null module's profile pointer
- after reload configuration, modules which will be removed have null profile pointer
- caused sigsegv after removing a module from the configuration

+++ supervisor.c
@@ -1821,7 +1821,9 @@ void service_stop_modules_sigint()
 {
    unsigned int x;
    for (x=0; x<loaded_modules_cnt; x++) {
-      if (running_modules[x].module_status == TRUE && (running_modules[x].modules_profile->profile_enabled == FALSE || running_modules[x].module_enabled == FALSE) && running_modules[x].sent_sigint == FALSE) {
+      if (running_modules[x].module_status == TRUE
+          && ((running_modules[x].modules_profile != NULL && running_modules[x].modules_profile->profile_enabled == FALSE) || running_modules[x].module_enabled == FALSE)
+          && running_modules[x].sent_sigint == FALSE) {
          #ifdef nemea_plugin
             netconf_notify(MODULE_EVENT_STOPPED,running_modules[x].module_name);
          #endif
@@ -1841,7 +1843,9 @@ void service_stop_modules_sigkill()
    unsigned int x, y;
 
    for (x = 0; x < loaded_modules_cnt; x++) {
-      if (running_modules[x].module_status == TRUE && (running_modules[x].module_enabled == FALSE || running_modules[x].modules_profile->profile_enabled == FALSE) && running_modules[x].sent_sigint == TRUE) {
+      if (running_modules[x].module_status == TRUE
+          && (running_modules[x].module_enabled == FALSE || (running_modules[x].modules_profile != NULL && running_modules[x].modules_profile->profile_enabled == FALSE))
+          && running_modules[x].sent_sigint == TRUE) {
          VERBOSE(MODULE_EVENT, "%s [STOP] Stopping module %s... sending SIGKILL\n", get_formatted_time(), running_modules[x].module_name);
          kill(running_modules[x].module_pid,9);
 
@@ -1893,13 +1897,18 @@ void service_update_modules_status()
          max_restarts = max_restarts_per_minute_config;
       }
 
-      if (running_modules[x].modules_profile->profile_enabled == TRUE && running_modules[x].module_enabled == TRUE && running_modules[x].module_status == FALSE && (running_modules[x].module_restart_cnt == max_restarts)) {
+      if ((running_modules[x].modules_profile != NULL && running_modules[x].modules_profile->profile_enabled == TRUE)
+           && running_modules[x].module_enabled == TRUE
+           && running_modules[x].module_status == FALSE
+           && (running_modules[x].module_restart_cnt == max_restarts)) {
          VERBOSE(MODULE_EVENT,"%s [RESTART] Module: %s was restarted %d times per minute and it is down again. I set it disabled.\n", get_formatted_time(), running_modules[x].module_name, max_restarts);
          running_modules[x].module_enabled = FALSE;
          #ifdef nemea_plugin
             netconf_notify(MODULE_EVENT_DISABLED,running_modules[x].module_name);
          #endif
-      } else if (running_modules[x].modules_profile->profile_enabled == TRUE && running_modules[x].module_status == FALSE && running_modules[x].module_enabled == TRUE) {
+      } else if ((running_modules[x].modules_profile != NULL && running_modules[x].modules_profile->profile_enabled == TRUE)
+                 && running_modules[x].module_status == FALSE
+                 && running_modules[x].module_enabled == TRUE) {
          service_start_module(x);
       }
    }
